satisfy clippy
